Move ajouter-cas-types/ to ir/
