+++ components/ir/ajouter-cas-types/ajouter-cas-types-component.jsx
