+++ components/ir/ajouter-cas-types/index.ts
@@ -5,11 +5,11 @@ import {
   createCasType,
   simulateCasTypes,
   updateCasType,
-} from "../../redux/actions";
+} from "../../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../redux/reducers";
+import { RootState } from "../../../redux/reducers";
 // eslint-disable-next-line no-unused-vars
-import { CasType } from "../../redux/reducers/descriptions/ir";
+import { CasType } from "../../../redux/reducers/descriptions/ir";
 import AjouterCasTypesComponent from "./ajouter-cas-types-component";
 
 const randomGender = () => (Math.random() < 0.49 ? 1 : 0);
